Move abstract_models into models package
